# examples/user-management/out/internal/api/handler_org_audit_events.go
// Code generated by stego. DO NOT EDIT.

package api

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"reflect"
	"strconv"
	"strings"

	"github.com/google/uuid"
)

// OrgAuditEventsHandler handles HTTP requests for AuditEvent entities.
type OrgAuditEventsHandler struct {
	store Storage
}

// NewOrgAuditEventsHandler creates a new OrgAuditEventsHandler.
func NewOrgAuditEventsHandler(store Storage) *OrgAuditEventsHandler {
	return &OrgAuditEventsHandler{store: store}
}

func (h *OrgAuditEventsHandler) Create(w http.ResponseWriter, r *http.Request) {
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	{
		var rawMap map[string]json.RawMessage
		if err := json.Unmarshal(bodyBytes, &rawMap); err == nil {
			if kindRaw, ok := rawMap["kind"]; ok {
				var kind string
				if err := json.Unmarshal(kindRaw, &kind); err != nil || kind != "AuditEvent" {
					handleError(w, r, BadRequest("kind must be AuditEvent"))
					return
				}
			}
		}
	}
	var auditevent AuditEvent
	if err := json.Unmarshal(bodyBytes, &auditevent); err != nil {
		handleError(w, r, BadRequest(err.Error()))
		return
	}
	auditevent.ID = uuid.New().String()
	auditevent.SourceType = "Organization"
	if err := h.store.Create(r.Context(), "AuditEvent", auditevent); err != nil {
		if errors.Is(err, ErrConflict) {
			handleError(w, r, Conflict("resource already exists"))
		} else {
			handleError(w, r, InternalError("internal error"))
		}
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(presentEntity(auditevent, "AuditEvent", auditevent.ID, "/api/user-mgmt/v1/org-events"+"/"+auditevent.ID))
}

func (h *OrgAuditEventsHandler) List(w http.ResponseWriter, r *http.Request) {
	pageStr := r.URL.Query().Get("page")
	pageSizeStr := r.URL.Query().Get("pageSize")
	page, _ := strconv.Atoi(pageStr)
	if page < 1 {
		page = 1
	}
	pageSize, _ := strconv.Atoi(pageSizeStr)
	if pageSize < 1 {
		pageSize = 100
	}
	if pageSize > 65500 {
		pageSize = 65500
	}
	validFields := map[string]bool{"source_type": true, "source_id": true, "action": true, "detail": true}
	orderDir := r.URL.Query().Get("order")
	if orderDir != "" && orderDir != "asc" && orderDir != "desc" {
		handleError(w, r, BadRequest("invalid order value: "+orderDir+"; must be 'asc' or 'desc'"))
		return
	}
	var orderBy []OrderByField
	if orderByStr := r.URL.Query().Get("orderBy"); orderByStr != "" {
		for _, entry := range strings.Split(orderByStr, ",") {
			entry = strings.TrimSpace(entry)
			if entry == "" {
				continue
			}
			parts := strings.Fields(entry)
			fieldName := parts[0]
			dir := "asc"
			if len(parts) > 1 {
				d := strings.ToLower(parts[1])
				if d != "asc" && d != "desc" {
					handleError(w, r, BadRequest("invalid orderBy direction: "+parts[1]+"; must be 'asc' or 'desc'"))
					return
				}
				dir = d
			}
			if !validFields[fieldName] {
				handleError(w, r, BadRequest("invalid orderBy field: "+fieldName))
				return
			}
			orderBy = append(orderBy, OrderByField{Field: fieldName, Direction: dir})
		}
	} else {
		orderBy = []OrderByField{{Field: "created_time", Direction: "desc"}}
	}
	if orderDir != "" {
		for i := range orderBy {
			orderBy[i].Direction = orderDir
		}
	}
	var fields []string
	if fieldsStr := r.URL.Query().Get("fields"); fieldsStr != "" {
		hasID := false
		for _, f := range strings.Split(fieldsStr, ",") {
			f = strings.TrimSpace(f)
			if f == "" {
				continue
			}
			if f != "id" && !validFields[f] {
				handleError(w, r, BadRequest("invalid fields value: "+f))
				return
			}
			fields = append(fields, f)
			if f == "id" {
				hasID = true
			}
		}
		if !hasID {
			fields = append(fields, "id")
		}
	}
	searchExpr := r.URL.Query().Get("search")
	opts := ListOptions{Page: page, Size: pageSize, OrderBy: orderBy, Fields: fields, Search: searchExpr, ImplicitFilters: map[string]string{"source_type": "Organization"}}
	listResult, err := h.store.List(r.Context(), "AuditEvent", "", "", opts)
	if err != nil {
		if errors.Is(err, ErrSearch) {
			handleError(w, r, BadRequest(err.Error()))
		} else {
			handleError(w, r, InternalError(err.Error()))
		}
		return
	}
	w.Header().Set("Content-Type", "application/json")
	hrefBase := "/api/user-mgmt/v1/org-events"
	itemsSlice := reflect.ValueOf(listResult.Items)
	actualSize := itemsSlice.Len()
	presentedItems := make([]map[string]any, actualSize)
	for i := 0; i < actualSize; i++ {
		item := itemsSlice.Index(i).Interface()
		itemID := reflect.ValueOf(item).FieldByName("ID").String()
		itemData, err := json.Marshal(item)
		if err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		var apiItem AuditEvent
		if err := json.Unmarshal(itemData, &apiItem); err != nil {
			handleError(w, r, InternalError("internal error"))
			return
		}
		presentedItems[i] = presentEntity(apiItem, "AuditEvent", itemID, hrefBase+"/"+itemID)
		if len(fields) > 0 {
			allowed := map[string]bool{"id": true, "kind": true, "href": true}
			for _, f := range fields {
				allowed[f] = true
			}
			for k := range presentedItems[i] {
				if !allowed[k] {
					delete(presentedItems[i], k)
				}
			}
		}
	}
	result := map[string]any{
		"kind":  "AuditEventList",
		"page":  page,
		"size":  actualSize,
		"total": listResult.Total,
		"items": presentedItems,
	}
	json.NewEncoder(w).Encode(result)
}