# examples/user-management/out/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Generated API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/user-mgmt/v1/org-events": {
      "get": {
        "summary": "List AuditEvent entities via org-audit-events",
        "operationId": "listOrgAuditEvents",
        "tags": [
          "org-audit-events"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "description": "Comma-separated list of field_name or field_name asc|desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "description": "Sort direction override: applied to all orderBy entries",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "description": "Comma-separated field names for sparse fieldset selection",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "TSL filter expression for searching entities",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/AuditEvent"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "href": {
                                "type": "string"
                              },
                              "id": {
                                "type": "string"
                              },
                              "kind": {
                                "type": "string"
                              }
                            }
                          }
                        ]
                      }
                    },
                    "kind": {
                      "type": "string"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create AuditEvent via org-audit-events",
        "operationId": "createOrgAuditEvents",
        "tags": [
          "org-audit-events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditEventCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AuditEvent"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "href": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditEvent": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 100
          },
          "detail": {
            "description": "Arbitrary JSON value"
          },
          "source_id": {
            "type": "string",
            "maxLength": 255
          },
          "source_type": {
            "type": "string",
            "maxLength": 50
          }
        },
        "required": [
          "source_type",
          "source_id",
          "action"
        ]
      },
      "AuditEventCreateRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 100
          },
          "detail": {
            "description": "Arbitrary JSON value"
          },
          "source_id": {
            "type": "string",
            "maxLength": 255
          },
          "source_type": {
            "type": "string",
            "maxLength": 50
          }
        },
        "required": [
          "source_type",
          "source_id",
          "action"
        ]
      },
...
    }
  }
}