# examples/user-management/out/internal/storage/models.go
// Code generated by stego. DO NOT EDIT.

package storage

import (
	"github.com/google/uuid"
	"gorm.io/datatypes"
	"gorm.io/gorm"
	"time"
)

// Meta is the base model, embedded in all entities.
type Meta struct {
	ID          string         `json:"id"`
	CreatedTime time.Time      `json:"created_time" gorm:"autoCreateTime"`
	UpdatedTime time.Time      `json:"updated_time" gorm:"autoUpdateTime"`
	DeletedAt   gorm.DeletedAt `json:"-" gorm:"index"`
}

// BeforeCreate is a GORM hook that auto-generates a UUID for ID on create.
func (m *Meta) BeforeCreate(tx *gorm.DB) error {
	if m.ID == "" {
		m.ID = uuid.New().String()
	}
	return nil
}

// Organization represents the Organization entity.
type Organization struct {
	Meta
	Name        string  `json:"name" gorm:"column:name;not null;uniqueIndex;size:255;check:length(name) >= 1"`
	Description *string `json:"description,omitempty" gorm:"column:description;size:1000"`
}

// User represents the User entity.
type User struct {
	Meta
	Email       string         `json:"email" gorm:"column:email;not null;uniqueIndex;size:255"`
	DisplayName string         `json:"display_name" gorm:"column:display_name;not null;size:100;check:length(display_name) >= 1"`
	Role        string         `json:"role" gorm:"column:role;not null"`
	OrgID       string         `json:"org_id" gorm:"column:org_id;not null"`
	OrgIDRef    *Organization  `json:"-" gorm:"foreignKey:OrgID"`
	Metadata    datatypes.JSON `json:"metadata,omitempty" gorm:"column:metadata;type:jsonb"`
}

// OrgSetting represents the OrgSetting entity.
type OrgSetting struct {
	Meta
	OrgID      string         `json:"org_id" gorm:"column:org_id;not null;uniqueIndex:upsert_org_id_key"`
	OrgIDRef   *Organization  `json:"-" gorm:"foreignKey:OrgID"`
	Key        string         `json:"key" gorm:"column:key;not null;size:255;uniqueIndex:upsert_org_id_key"`
	Value      datatypes.JSON `json:"value" gorm:"column:value;type:jsonb;not null"`
	Generation int64          `json:"generation" gorm:"column:generation;not null"`
}

// AuditEvent represents the AuditEvent entity.
type AuditEvent struct {
	Meta
	SourceType string         `json:"source_type" gorm:"column:source_type;not null;size:50"`
	SourceID   string         `json:"source_id" gorm:"column:source_id;not null;size:255"`
	Action     string         `json:"action" gorm:"column:action;not null;size:100"`
	Detail     datatypes.JSON `json:"detail,omitempty" gorm:"column:detail;type:jsonb"`
}